mandatory events callback examples combined
