+++ examples/ompt/CMakeLists.txt
@@ -1,7 +1,5 @@
 if(HPXMP_WITH_OMPT)
     set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -std=c11 -fopenmp")
     add_executable(hello_world hello_world.c)
-    add_executable(task_create task_create.c)
-    add_executable(task_schedule task_schedule.c)
-    add_executable(implicit_task implicit_task.c)
+    add_executable(task task.c)
 endif()
